fix: separate out config from the command
Cogs seem to duplicate the module, separately from the usual effect of
importing it. I will investigate further, but this does fix it.
